Splitting parsers into their own module
