+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod matchspec;
+mod parsers;
 pub use crate::matchspec::*;
